Add event listeners to buttons and container div
I struggled with adding the addEventListeners to the 3 buttons. In the
end I used the for of loop to add to each one. Then I erased and
replaced some of the code from the first time which used a lot of
console.logs. I forgot to stop to this step and I started to add the div
container for the result div to be appended.